+++ src/index.ts
@@ -23,6 +23,8 @@ export { ModalBody } from './components/ui/Modal/ModalBody';
 export { ModalButton } from './components/ui/Modal/ModalButton';
 export { ModalButtonGroup } from './components/ui/Modal/ModalButtonGroup';
 export { ModalHeader } from './components/ui/Modal/ModalHeader';
+export { Navbar } from './components/ui/Navbar';
+export { NavbarItem } from './components/ui/Navbar/NavbarItem';
 export { Notification } from './components/ui/Notification';
 export { NotificationGroup } from './components/ui/NotificationGroup';
 export { PopOver } from './components/ui/PopOver';